normal bound objects transfered with cef ipc as well
