+++ CefSharp.Core/CefSharp.Core.vcxproj.filters
@@ -74,6 +74,9 @@
     <ClCompile Include="Internals\CookieManager.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
+    <ClCompile Include="Internals\Serialization\JsObjectsSerialization.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
     <ClCompile Include="Internals\Serialization\ObjectsSerialization.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
@@ -226,10 +229,13 @@
     <ClInclude Include="RequestContextSettings.h">
       <Filter>Header Files</Filter>
     </ClInclude>
-    <ClInclude Include="Internals\Serialization\ObjectsSerialization.h">
+    <ClInclude Include="Internals\CookieManager.h">
       <Filter>Header Files</Filter>
     </ClInclude>
-    <ClInclude Include="Internals\CookieManager.h">
+    <ClInclude Include="Internals\Serialization\JsObjectsSerialization.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
+    <ClInclude Include="Internals\Serialization\ObjectsSerialization.h">
       <Filter>Header Files</Filter>
     </ClInclude>
   </ItemGroup>
