+++ CefSharp.Core/Internals/Serialization/V8Serialization.cpp
@@ -18,70 +18,6 @@ namespace CefSharp
             void SerializeV8SimpleObject(Object^ obj, CefRefPtr<TList> list, TIndex index, Stack<Object^>^ seen);
 
             template<typename TList, typename TIndex>
-            Object^ DeserializeV8Object(CefRefPtr<TList> list, TIndex index, IJavascriptCallbackFactory^ javascriptCallbackFactory)
-            {
-                Object^ result = nullptr;
-                auto type = list->GetType(index);
-
-                if (type == VTYPE_BOOL)
-                {
-                    result = list->GetBool(index);
-                }
-                else if (type == VTYPE_INT)
-                {
-                    result = list->GetInt(index);
-                }
-                else if (IsInt64(list, index))
-                {
-                    result = GetInt64(list, index);
-                }
-                else if (IsCefTime(list, index))
-                {
-                    auto cefTime = GetCefTime(list, index);
-                    result = ConvertCefTimeToDateTime(cefTime);
-                }
-                else if (IsJsCallback(list, index) && javascriptCallbackFactory != nullptr)
-                {
-                    auto jsCallbackDto = GetJsCallback(list, index);
-                    result = javascriptCallbackFactory->Create(jsCallbackDto);
-                }
-                else if (type == VTYPE_DOUBLE)
-                {
-                    result = list->GetDouble(index);
-                }
-                else if (type == VTYPE_STRING)
-                {
-                    result = StringUtils::ToClr(list->GetString(index));
-                }
-                else if (type == VTYPE_LIST)
-                {
-                    auto subList = list->GetList(index);
-                    auto array = gcnew List<Object^>(subList->GetSize());
-                    for (auto i = 0; i < subList->GetSize(); i++)
-                    {
-                        array->Add(DeserializeV8Object(subList, i, javascriptCallbackFactory));
-                    }
-                    result = array;
-                }
-                else if (type == VTYPE_DICTIONARY)
-                {
-                    auto dict = gcnew Dictionary<String^, Object^>();
-                    auto subDict = list->GetDictionary(index);
-                    std::vector<CefString> keys;
-                    subDict->GetKeys(keys);
-
-                    for (auto i = 0; i < keys.size(); i++)
-                    {
-                        dict->Add(StringUtils::ToClr(keys[i]), DeserializeV8Object(subDict, keys[i], javascriptCallbackFactory));
-                    }
-
-                    result = dict;
-                }
-
-                return result;
-            }
-
-            template<typename TList, typename TIndex>
             void SerializeV8Object(Object^ obj, CefRefPtr<TList> list, TIndex index)
             {
                 auto seen = gcnew Stack<Object^>();
@@ -166,16 +102,6 @@ namespace CefSharp
                 seen->Pop();
             }
 
-            DateTime ConvertCefTimeToDateTime(CefTime time)
-            {
-                auto epoch = time.GetDoubleT();
-                if (epoch == 0)
-                {
-                    return DateTime::MinValue;
-                }
-                return DateTime(1970, 1, 1, 0, 0, 0).AddSeconds(epoch).ToLocalTime();
-            }
-
             CefTime ConvertDateTimeToCefTime(DateTime dateTime)
             {
                 auto timeSpan = dateTime - DateTime(1970, 1, 1);
@@ -183,8 +109,6 @@ namespace CefSharp
                 return CefTime(timeSpan.TotalSeconds);
             }
 
-            template Object^ DeserializeV8Object(CefRefPtr<CefListValue> list, int index, IJavascriptCallbackFactory^ javascriptCallbackFactory);
-            template Object^ DeserializeV8Object(CefRefPtr<CefDictionaryValue> list, CefString index, IJavascriptCallbackFactory^ javascriptCallbackFactory);
             template void SerializeV8Object(Object^ obj, CefRefPtr<CefListValue> list, int index);
             template void SerializeV8Object(Object^ obj, CefRefPtr<CefDictionaryValue> list, CefString index);
         }
